Refactored profile transformation.
