+++ src/psyclone/tests/profile_test.py
@@ -46,8 +46,8 @@ from psyclone.generator import GenerationError
 from psyclone.profiler import Profiler, ProfileNode
 from psyclone.psyGen import Loop, NameSpace, InternalError
 from psyclone.psyir.transformations import TransformationError
-from psyclone.transformations import GOceanOMPLoopTrans, OMPParallelTrans, \
-    ProfileRegionTrans
+from psyclone.psyir.transformations import ProfileRegion
+from psyclone.transformations import GOceanOMPLoopTrans, OMPParallelTrans
 from psyclone.tests.utilities import get_invoke
 
 
@@ -110,7 +110,7 @@ def test_profile_basic(capsys):
         "it_space='go_internal_pts']\n")
     assert expected in out
 
-    prt = ProfileRegionTrans()
+    prt = ProfileRegion()
 
     # Insert a profile call between outer and inner loop.
     # This tests that we find the subroutine node even
@@ -474,9 +474,9 @@ def test_transform(capsys):
                            name="invoke_loop1")
     schedule = invoke.schedule
 
-    prt = ProfileRegionTrans()
+    prt = ProfileRegion()
     assert str(prt) == "Insert a profile start and end call."
-    assert prt.name == "ProfileRegionTrans"
+    assert prt.name == "ProfileRegion"
 
     # Try applying it to a list
     sched1, _ = prt.apply(schedule.children)
@@ -629,7 +629,7 @@ def test_transform_errors(capsys):
                            name="invoke_loop1")
 
     schedule = invoke.schedule
-    prt = ProfileRegionTrans()
+    prt = ProfileRegion()
 
     with pytest.raises(TransformationError) as excinfo:
         prt.apply([schedule.children[0].children[0], schedule.children[1]])
@@ -694,7 +694,7 @@ def test_transform_errors(capsys):
                            name="invoke_loop1")
     schedule = invoke.schedule
 
-    prt = ProfileRegionTrans()
+    prt = ProfileRegion()
     omp_loop = GOceanOMPLoopTrans()
 
     # Parallelise the first loop:
@@ -718,7 +718,7 @@ def test_omp_transform():
                            name="invoke_loop1")
     schedule = invoke.schedule
 
-    prt = ProfileRegionTrans()
+    prt = ProfileRegion()
     omp_loop = GOceanOMPLoopTrans()
     omp_par = OMPParallelTrans()
