+++ src/psyclone/transformations.py
@@ -2649,161 +2649,6 @@ class OCLTrans(Transformation):
                     format(kern.name, [sym.name for sym in global_variables]))
 
 
-class ProfileRegionTrans(RegionTrans):
-    ''' Create a profile region around a list of statements. For
-    example:
-
-    >>> from psyclone.parse.algorithm import parse
-    >>> from psyclone.parse.utils import ParseError
-    >>> from psyclone.psyGen import PSyFactory, GenerationError
-    >>> api = "gocean1.0"
-    >>> filename = "nemolite2d_alg.f90"
-    >>> ast, invokeInfo = parse(filename, api=api, invoke_name="invoke")
-    >>> psy = PSyFactory(api).create(invokeInfo)
-    >>>
-    >>> from psyclone.psyGen import TransInfo
-    >>> t = TransInfo()
-    >>> p_trans = t.get_trans_name('ProfileRegionTrans')
-    >>>
-    >>> schedule = psy.invokes.get('invoke_0').schedule
-    >>> schedule.view()
-    >>>
-    >>> # Enclose all children within a single profile region
-    >>> newschedule, _ = p_trans.apply(schedule.children)
-    >>> newschedule.view()
-
-    '''
-    from psyclone import psyGen, profiler
-    # Unlike other transformations we can be fairly relaxed about the nodes
-    # that a region can contain as we don't have to understand them.
-    valid_node_types = (psyGen.Node,)
-
-    def __str__(self):
-        return "Insert a profile start and end call."
-
-    @property
-    def name(self):
-        ''' Returns the name of this transformation as a string '''
-        return "ProfileRegionTrans"
-
-    def validate(self, nodes, options=None):
-        '''
-        Calls the validate method of the base class and then checks that,
-        for the NEMO API, the routine that will contain the profiling
-        region already has a Specification_Part (because we've not yet
-        implemented the necessary support if it doesn't).
-
-        :param nodes: a list of nodes to be profiled.
-        :type nodes: :py:class:`psyclone.psyGen.Loop`
-
-        :param options: a dictionary with options for transformations.
-        :type options: dictionary of string:values or None
-
-        :raises TransformationError: if we're using the NEMO API and the \
-                                     target routine has no Specification_Part.
-        '''
-        from fparser.two import Fortran2003
-        from fparser.two.utils import walk_ast
-        from psyclone.nemo import NemoInvoke
-
-        super(ProfileRegionTrans, self).validate(nodes, options)
-
-        # The checks below are only for the NEMO API and can be removed
-        # once #435 is done.
-        invoke = nodes[0].root.invoke
-        if not isinstance(invoke, NemoInvoke):
-            return
-        # Get the parse tree of the routine containing this region
-        ptree = invoke._ast
-        # Search for the Specification_Part
-        if not walk_ast([ptree], [Fortran2003.Specification_Part]):
-            raise TransformationError(
-                "For the NEMO API, profiling can only be added to routines "
-                "which contain existing variable declarations (i.e. a "
-                "Specification Part) but '{0}' does not have any.".format(
-                    invoke.name))
-
-    def apply(self, nodes, options=None):
-        # pylint: disable=arguments-differ
-        '''Apply this transformation to a subset of the nodes within a
-        schedule - i.e. enclose the specified Nodes in the
-        schedule within a single profiler region.
-
-        :param nodes: can be a single node or a list of nodes.
-        :type nodes: :py:obj:`psyclone.psygen.Node` or list of\
-                     :py:obj:`psyclone.psygen.Node`
-        :param options: a dictionary with options for transformations.
-        :type options: dictionary of string:values or None
-
-        :returns: Tuple of the modified schedule and a record of the \
-                  transformation.
-        :rtype: (:py:class:`psyclone.psyGen.Schedule`, \
-                :py:class:`psyclone.undoredo.Memento`)
-
-        '''
-        # Check whether we've been passed a list of nodes or just a
-        # single node.
-        from psyclone.psyGen import Node, OMPDoDirective, ACCLoopDirective
-        if isinstance(nodes, list) and isinstance(nodes[0], Node):
-            node_list = nodes
-        elif isinstance(nodes, Schedule):
-            # We've been passed a Schedule so default to enclosing its
-            # children.
-            node_list = nodes.children
-        elif isinstance(nodes, Node):
-            # Single node that's not a Schedule
-            node_list = [nodes]
-        else:
-            arg_type = str(type(nodes))
-            raise TransformationError("Error in {1}. "
-                                      "Argument must be a single Node in a "
-                                      "schedule or a list of Nodes in a "
-                                      "schedule but have been passed an "
-                                      "object of type: {0}".
-                                      format(arg_type, str(self)))
-
-        # Keep a reference to the parent of the nodes that are to be
-        # enclosed within a profile region. Also keep the index of
-        # the first child to be enclosed as that will become the
-        # position of the new Profile node
-        node_parent = node_list[0].parent
-        if isinstance(node_parent, Schedule) and \
-           isinstance(node_parent.parent, (OMPDoDirective, ACCLoopDirective)):
-            raise TransformationError("A ProfileNode cannot be inserted "
-                                      "between an OpenMP/ACC directive and "
-                                      "the loop(s) to which it applies!")
-        node_position = node_list[0].position
-
-        # Perform validation checks
-        self.validate(node_list, options)
-
-        # create a memento of the schedule and the proposed
-        # transformation
-        schedule = node_list[0].root
-
-        keep = Memento(schedule, self)
-
-        # Create the ProfileNode. All of the supplied child nodes will have
-        # the Profile's Schedule as their parent.
-        from psyclone.profiler import ProfileNode
-        profile_node = ProfileNode(parent=node_parent, children=node_list[:])
-
-        # Correct the parent's list of children. Use a slice of the list of
-        # nodes so that we're looping over a local copy of the list. Otherwise
-        # things get confused when we remove children from the list.
-        for child in node_list[:]:
-            # Remove child from the parent's list of children
-            node_parent.children.remove(child)
-
-        # Add the Profile node as a child of the parent
-        # of the nodes being enclosed and at the original location
-        # of the first of these nodes
-        node_parent.addchild(profile_node,
-                             index=node_position)
-
-        return schedule, keep
-
-
 class Dynamo0p3AsyncHaloExchangeTrans(Transformation):
     '''Splits a synchronous halo exchange into a halo exchange start and
     halo exchange end. For example:
